# docs/OPENAPI.yaml
openapi: 3.0.3
info:
  title: FlashDB HTTP API
  version: 1.0.0
  description: |
    Versioned HTTP API for FlashDB administrative operations.
servers:
  - url: http://localhost:8080
paths:
  /api/v1/execute:
    post:
      summary: Execute a FlashDB command
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommandRequest'
      responses:
        '200':
          description: Command response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandResponse'
...
  /api/v1/keys:
    get:
      summary: List keys
      parameters:
        - in: query
          name: pattern
          schema:
            type: string
            default: '*'
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            default: 100
      responses:
        '200':
          description: Keys list
          content:
            application/json:
              schema:
                type: object
                properties:
                  keys:
                    type: array
                    items:
                      $ref: '#/components/schemas/KeyInfo'
                  total:
                    type: integer
                required: [keys, total]
...
components:
  schemas:
    CommandRequest:
      type: object
      properties:
        command:
          type: string
        args:
          type: array
          items:
            type: string
      required: [command]
    CommandResponse:
      type: object
      properties:
        success:
          type: boolean
        result:
          nullable: true
        error:
          type: string
        type:
          type: string
      required: [success]
...
    KeyInfo:
      type: object
      properties:
        key:
          type: string
        type:
          type: string
        ttl:
          type: integer
        value:
          type: string
      required: [key, type, ttl]
...